Changed SecureStoreData and SecureStoreMetadata to classes
